+++ app/models/referee.rb
@@ -14,6 +14,7 @@ class Referee < ApplicationRecord
   has_many :referee_taggings, dependent: :destroy
   has_many :referee_club_exclusions, dependent: :destroy
   has_many :referee_club_exclusion_requests, dependent: :destroy
+  has_many :referee_change_requests, dependent: :destroy
   has_many :game_day_referee_confirmations, dependent: :destroy
   has_many :referee_qualification_types, through: :referee_qualifications
   has_many :referee_tags, through: :referee_taggings